Add the version of genqlient to the generated code
Craig suggested this; it seems plausibly useful although of course it
will cause some churn in the generated file as well.  And it's easy to
do from Go's automatically added module metadata.  Luckily, something
about how that works means that even when you're on a locally-tagged
commit, in tests the version is `(devel)`, so hopefully this won't cause
any problems for them.

Test plan:
make check

Reviewers: steve