Add `omitempty` for pointer input fields
When `optional: "pointer"` is configured, `omitempty` is needed for input
fields for the generated code to be semantically correct. Without `omitempty`
an empty pointer would generate an "empty value" which is incorrect.

Consider the following graphql schema generated by Hasura:

```
input task_insert_input {
  id: Int
}
```

Without this fix, the following golang struct will be generated:
```
type Task_insert_input struct {
  Id: *int `json:"id"`
}
```

When we populate the struct without `Id` being set, it'd serialize to
`{ id: 0 }` as oppose to `{}` (i.e. so that the id would be generated by Hasura)