Add some better tests for use_struct_references
The test it was using was not really exercising things (maybe this was
before I added support for using different queries for config tests).
The new tests exercise a bit more, which is useful if you want to know
exactly what this option does. I also added tests where we mix with
`optional: pointer` since those interact a bit.