Reorganize and improve documentation
We've been in need of a documentation revamp for a while; a giant FAQ
was never the greatest structure and got worse as it grew. In this
commit I reorganize the documentation. Most of it is just moving around
existing text, but I added some new documentation here and there.

The changes:
- Many of the FAQ questions have moved to several new docs, on the
  client/runtime, handling your schema, and various operation types; the
  FAQ has answers to a few of the actually most frequent questions, as
  well as a few things that didn't fit elsewhere.
- We now have a `docs/README.md` which acts as an index, so we can just
  link to `/docs`.
- I lowercased the files that don't need match a GitHub convention, so
  it's a bit less yell-y.
- I added documentation on:
  - how we version genqlient (fixes #63)
  - newer options for optional types
  - a bit more on custom scalars and integer types (fixes #128)