disable input validation when `use_struct_references` is enabled (#344)
Fix for https://github.com/Khan/genqlient/issues/342. As
`use_struct_references` is incompatible with `omitempty` and `pointer`
validation, the most simple change is to disable the validation when the
`use_struct_references` is enabled. Note this implementation disables
the validation even for types that are not influenced by
`use_struct_references`, but still better than disabling the validation
globally.

I have:
- [x] Written a clear PR title and description (above)
- [x] Signed the [Khan Academy CLA](https://www.khanacademy.org/r/cla)
- [x] Added tests covering my changes, if applicable
- [x] Included a link to the issue fixed, if applicable
- [ ] Included documentation, for new features
- [x] Added an entry to the changelog