Remove allow_broken_features
This was a config option from the early days, when there were various
things I wanted to be able to try but didn't really want users to use
yet. With subscriptions merged, the last plausible need is long gone!

Technically this is a breaking change, but it was documented as unstable
so I didn't mark it as one.