Subscription connection parameters support (#360)
Adds support for subscription (websocket) connection parameters support.

We are having an issue with authorization when using subscriptions in
genqlient. Our mobile app team is using Apollo for GraphQL requests, and
Apollo seems to sends its auth in a way that is different from how
genqlient does it:


https://www.apollographql.com/docs/react/data/subscriptions#5-authenticate-over-websocket-optional

This is causing a problems, because we use genqlient for testing the
backend. It seems that the way Apollo is implemented, authentication
(and other headers) are not actually sent as HTTP headers, but as
connection parameters instead.

I have:
- [x] Written a clear PR title and description (above)
- [x] Signed the [Khan Academy CLA](https://www.khanacademy.org/r/cla)
- [x] Added tests covering my changes, if applicable
- [x] Included a link to the issue fixed, if applicable
- [x] Included documentation, for new features
- [x] Added an entry to the changelog

---------

Co-authored-by: Ben Kraft <ben@benkraft.org>