Extend auto_camel_case to type names
This extends the auto_camel_case feature to apply to GraphQL type names in addition to field names. With this change, GraphQL types using snake_case (like 'user_profile') will be automatically converted to camelCase and then capitalized for Go (like 'UserProfile').

The feature works by:
1. Extending the existing auto_camel_case functionality to type names
2. Adding the snakeToCamel conversion to type names in various places:
   - In typeNameParts to handle nested type paths
   - In InputObject and Enum type name generation
3. Updating documentation and adding test cases with snake_case type names

The original type names are preserved in the GraphQL query for server compatibility.