Fix test flake revealed by #308 (#393)
Two issues here:
- In the with-config tests, we were generating all the different input
files into one output. This makes sense in user code, but in our tests,
we reuse the same types all over the place, so not only is it confusing
to read all the tests interleaved, it can cause nondeterminism anytime
our conflict-detection is incomplete (#123). As it happens, #308
introduced such a case, so tests started flaking. Fixes #281.
- This turned the case introduced by #308 from a flake into a consistent
failure. The issue is actually unrelated to the changes in the PR (it's
intentional validation added in #338, and specifically corresponds to
`errors/OmitemptyDirective.graphql` from that PR), so I changed the
config for the test case to one that avoids the issue while still
testing what we want.